SqlEventPublisher.PublishPendingEvents<>() 메서드가 지연된 이벤트가 없을 경우 메시지 발행을 시도하지 않도록합니다.

+++ source/Khala.EventSourcing.Sql/EventSourcing/Sql/SqlEventPublisher.cs
@@ -71,6 +71,11 @@
                     .Select(e => RestoreEnvelope(e))
                     .ToList();
 
+                if (envelopes.Any() == false)
+                {
+                    return;
+                }
+
                 await _messageBus.SendBatch(envelopes, cancellationToken).ConfigureAwait(false);
 
                 context.PendingEvents.RemoveRange(pendingEvents);
